Build in compatibility with the old pretty API
